+++ chocolatey/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.0 : 2025-04-30
+
+### New
+- Add tool for adding a package source.
+- Add tool for removing a package source.
+
 ## 1.0.0 : 2025-04-29
 
 ### Initial Version